% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{combine_plots}
\alias{combine_plots}
\title{Create a Publication-Ready Figure Grid}
\usage{
combine_plots(..., ncol = 2, heights = NULL)
}
\arguments{
\item{...}{ggplot objects to combine}

\item{ncol}{Number of columns in grid}

\item{heights}{Relative heights of rows (passed to patchwork)}
}
\value{
patchwork object
}
\description{
Combines multiple plots into a grid with consistent spacing and legend handling.
}
\examples{
\dontrun{
  p1 <- ggplot(mtcars, aes(x = mpg)) + geom_histogram()
  p2 <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
  combine_plots(p1, p2, ncol = 2)
}

}
